predictions/requirements.txt
google-cloud-storage
polling